Move LedgerTime and poll_beta_has_expired out of swap/action.rs
